lastUpdate

+++ lib/src/weather/data/model/weatherModel.dart
@@ -85,10 +85,10 @@ class MainModel extends Main {
     Map<String, dynamic> json,
   ) {
     return MainModel(
-      temp: json['temp'],
-      feelsLike: json['feels_like'],
-      tempMin: json['temp_min'],
-      tempMax: json['temp_max'],
+      temp: json['temp'].toDouble(),
+      feelsLike: json['feels_like'].toDouble(),
+      tempMin: json['temp_min'].toDouble(),
+      tempMax: json['temp_max'].toDouble(),
       pressure: json['pressure'].toDouble(),
       humidity: json['humidity'].toDouble(),
     );
